requirements.txt
appdirs==1.4.3
packaging==16.8
pkg-resources==0.0.0
PyGithub==1.33
PyJWT==1.4.2
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
